+++ sdk/digitaltwins/Azure.DigitalTwins.Core/src/Customized/DigitalTwinModelsRestClient.cs
@@ -88,7 +88,7 @@ namespace Azure.DigitalTwins.Core
 
         // The modelUpdates parameter needs to be changed from IEnumerable<object> to IEnumerable<string>
         // and not parsed like a json object.
-        internal async Task<Response> UpdateAsync(string id, IEnumerable<string> modelUpdates, UpdateModelOptions digitalTwinModelsUpdateOptions = null, CancellationToken cancellationToken = default)
+        internal async Task<Response> UpdateAsync(string id, IEnumerable<string> modelUpdates, DecomissionModelOptions digitalTwinModelsUpdateOptions = null, CancellationToken cancellationToken = default)
         {
             if (id == null)
             {
@@ -120,7 +120,7 @@ namespace Azure.DigitalTwins.Core
 
         // The modelUpdates parameter needs to be changed from IEnumerable<object> to IEnumerable<string>
         // and not parsed like a json object.
-        internal Response Update(string id, IEnumerable<string> modelUpdates, UpdateModelOptions digitalTwinModelsUpdateOptions = null, CancellationToken cancellationToken = default)
+        internal Response Update(string id, IEnumerable<string> modelUpdates, DecomissionModelOptions digitalTwinModelsUpdateOptions = null, CancellationToken cancellationToken = default)
         {
             if (id == null)
             {
@@ -162,13 +162,13 @@ namespace Azure.DigitalTwins.Core
             uri.AppendPath("/models", false);
             uri.AppendQuery("api-version", apiVersion, true);
             request.Uri = uri;
-            if (digitalTwinModelsAddOptions?.Traceparent != null)
+            if (digitalTwinModelsAddOptions?.TraceParent != null)
             {
-                request.Headers.Add("traceparent", digitalTwinModelsAddOptions.Traceparent);
+                request.Headers.Add("TraceParent", digitalTwinModelsAddOptions.TraceParent);
             }
-            if (digitalTwinModelsAddOptions?.Tracestate != null)
+            if (digitalTwinModelsAddOptions?.TraceState != null)
             {
-                request.Headers.Add("tracestate", digitalTwinModelsAddOptions.Tracestate);
+                request.Headers.Add("TraceState", digitalTwinModelsAddOptions.TraceState);
             }
             request.Headers.Add("Content-Type", "application/json");
             request.Headers.Add("Accept", "application/json");
@@ -182,7 +182,7 @@ namespace Azure.DigitalTwins.Core
 
         // The strings are already json, so we do not want them to be serialized.
         // Instead, the payloads need to be concatenated into a json array.
-        private HttpMessage CreateUpdateRequest(string id, IEnumerable<string> modelUpdates, UpdateModelOptions digitalTwinModelsUpdateOptions)
+        private HttpMessage CreateUpdateRequest(string id, IEnumerable<string> modelUpdates, DecomissionModelOptions digitalTwinModelsUpdateOptions)
         {
             var message = _pipeline.CreateMessage();
             var request = message.Request;
@@ -193,13 +193,13 @@ namespace Azure.DigitalTwins.Core
             uri.AppendPath(id, true);
             uri.AppendQuery("api-version", apiVersion, true);
             request.Uri = uri;
-            if (digitalTwinModelsUpdateOptions?.Traceparent != null)
+            if (digitalTwinModelsUpdateOptions?.TraceParent != null)
             {
-                request.Headers.Add("traceparent", digitalTwinModelsUpdateOptions.Traceparent);
+                request.Headers.Add("TraceParent", digitalTwinModelsUpdateOptions.TraceParent);
             }
-            if (digitalTwinModelsUpdateOptions?.Tracestate != null)
+            if (digitalTwinModelsUpdateOptions?.TraceState != null)
             {
-                request.Headers.Add("tracestate", digitalTwinModelsUpdateOptions.Tracestate);
+                request.Headers.Add("TraceState", digitalTwinModelsUpdateOptions.TraceState);
             }
             request.Headers.Add("Content-Type", "application/json-patch+json");
             request.Headers.Add("Accept", "application/json");