+++ sdk/digitaltwins/Azure.DigitalTwins.Core/src/Customized/Models/GetRelationshipsOptions.cs
@@ -10,5 +10,13 @@ namespace Azure.DigitalTwins.Core
     public partial class GetRelationshipsOptions
     {
         // This class declaration changes the namespace; do not remove.
+
+        /// <summary> Identifies the request in a distributed tracing system. </summary>
+        [CodeGenMember("Traceparent")]
+        public string TraceParent { get; set; }
+
+        /// <summary> Provides vendor-specific trace identification information and is a companion to TraceParent. </summary>
+        [CodeGenMember("Tracestate")]
+        public string TraceState { get; set; }
     }
 }