+++ sdk/digitaltwins/Azure.DigitalTwins.Core/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 - Regenerate protocol layer from service API version 2020-10-31
 - Update service API version to use service API version 2020-10-31 by default
-- Add optional parameters for traceparent and tracestate to all service request APIs to support distributed tracing
+- Add optional parameters for TraceParent and TraceState to all service request APIs to support distributed tracing
 - Renamed model type "ModelData" to "DigitalTwinsModelData" to make type less generic, and less likely to conflict with other libraries
 - Renamed model type "RequestOptions" to "DigitalTwinsRequestOptions" to make type less generic, and less likely to conflict with other libraries